refactor: move DatacollectorPermission to a dedicated permissions.py file
